Use the USE1 signal to retire a child
An new method has been introduced to split the stopping and the retiring of a
worker. Why is it necceassary? In the one hand when we are stopping a daemond
we do not want to wait long time. In the other hand we do not want to break
users' requests during code reloading.